# service/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"PrefixCode_ObjectDotNotation_BaseCode__": {
				"type": "string",
				"enum": [
					"server:success",
					"server:error",
					"server:auth.invalid_password",
					"server:auth.no_users",
					"server:auth.exists_email",
					"server:database.undefined",
					"server:database.duplicated",
					"server:database.disconnected"
				]
			},
			"BaseResponse_boolean.undefined_": {
				"properties": {
					"metadata": {},
					"message": {
						"type": "string"
					},
					"result": {
						"type": "boolean"
					},
					"code": {
						"$ref": "#/components/schemas/PrefixCode_ObjectDotNotation_BaseCode__"
					}
				},
				"required": [
					"result",
					"code"
				],
				"type": "object"
			},
			"UserMetaJson": {
				"properties": {
					"test": {
						"type": "string"
					}
				},
				"required": [
					"test"
				],
				"type": "object"
			},
			"UserSchema": {
				"properties": {
					"created_time": {
						"type": "number",
						"format": "double"
					},
					"updated_time": {
						"type": "number",
						"format": "double"
					},
					"meta_json": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/UserMetaJson"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"pwd": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"idx": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"created_time",
					"updated_time",
					"meta_json",
					"name",
					"pwd",
					"email",
					"type",
					"idx"
				],
				"type": "object"
			},
			"BaseResponse_UserSchema-Array._total-number__": {
				"properties": {
					"metadata": {
						"properties": {
							"total": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"total"
						],
						"type": "object"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/UserSchema"
						},
						"type": "array"
					},
					"code": {
						"$ref": "#/components/schemas/PrefixCode_ObjectDotNotation_BaseCode__"
					}
				},
				"required": [
					"result",
					"code"
				],
				"type": "object"
			},
			"BaseResponse__deleteCnt-number_.undefined_": {
				"properties": {
					"metadata": {},
					"message": {
						"type": "string"
					},
					"result": {
						"properties": {
							"deleteCnt": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"deleteCnt"
						],
						"type": "object"
					},
					"code": {
						"$ref": "#/components/schemas/PrefixCode_ObjectDotNotation_BaseCode__"
					}
				},
				"required": [
					"result",
					"code"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "service",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user/profile": {
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_boolean.undefined_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"properties": {},
											"additionalProperties": {
												"properties": {
													"message": {
														"type": "string"
													},
													"value": {
														"type": "string"
													}
												},
												"required": [
													"message",
													"value"
												],
												"type": "object"
											},
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Unhandled",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/list": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_UserSchema-Array._total-number__"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"properties": {},
											"additionalProperties": {
												"properties": {
													"message": {
														"type": "string"
													},
													"value": {
														"type": "string"
													}
												},
												"required": [
													"message",
													"value"
												],
												"type": "object"
											},
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Unhandled",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/{idx}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse__deleteCnt-number_.undefined_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"properties": {},
											"additionalProperties": {
												"properties": {
													"message": {
														"type": "string"
													},
													"value": {
														"type": "string"
													}
												},
												"required": [
													"message",
													"value"
												],
												"type": "object"
											},
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Unhandled",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idx",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/service/v1"
		}
	]
}